# trabaio/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API do Sistema de Petshop",
    "version": "1.0.0",
    "description": "API para gerenciamento de petshop com autenticação JWT"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AdminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/solicitations": {
      "get": {
        "summary": "Listar todas as solicitações",
        "tags": ["Solicitações"],
        "security": [
          {"BearerAuth": []}
        ],
        "responses": {
          "200": {
            "description": "Lista de solicitações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solicitation"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar nova solicitação",
        "tags": ["Solicitações"],
        "security": [
          {"BearerAuth": []}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitationCreation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Solicitação criada com sucesso"
          },
          "401": {
            "description": "Informação do usuário não encontrada no token."
          },
          "500": {
            "description": "Erro ao criar a solicitação no banco de dados."
          }
        }
      }
    },
    "/solicitations/{id}": {
      "get": {
        "summary": "Buscar solicitação por ID",
        "tags": ["Solicitações"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID da solicitação"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da solicitação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              }
            }
          },
          "404": {"description": "Solicitação não encontrada"},
          "500": {"description": "Erro ao buscar a solicitação"}
        }
      },
      "put": {
        "summary": "Atualizar solicitação por ID (completo)",
        "tags": ["Solicitações"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID da solicitação"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitationCreation"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Solicitação atualizada com sucesso"},
          "404": {"description": "Solicitação não encontrada"},
          "500": {"description": "Erro ao atualizar a solicitação"}
        }
      },
      "patch": {
        "summary": "Atualizar solicitação por ID (parcial)",
        "tags": ["Solicitações"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID da solicitação"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitationCreationPartial"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Solicitação atualizada parcialmente com sucesso"},
          "404": {"description": "Solicitação não encontrada"},
          "500": {"description": "Erro ao atualizar a solicitação parcialmente"},
          "delete": {
            "summary": "Deletar solicitação por ID",
            "tags": ["Solicitações"],
            "security": [
              {"BearerAuth": []}
            ],
            "parameters": [
              {
                "in": "path",
                "name": "id",
                "required": true,
                "schema": {"type": "integer"},
                "description": "ID da solicitação"
              }
            ],
            "responses": {
              "200": {"description": "Solicitação deletada com sucesso"},
              "404": {"description": "Erro ao deletar o serviço"},
              "500": {"description": "Erro ao deletar a solicitação"}
            }
          }
        }
      }
    },
    "/tutors": {
      "get": {
        "summary": "Listar todos os tutores",
        "tags": ["Tutores"],
        "responses": {
          "200": {
            "description": "Lista de tutores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tutor"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar novo tutor",
        "tags": ["Tutores"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorCreation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tutor criado com sucesso"
          },
          "500": {
            "description": "Erro ao criar tutor"
          }
        }
      },
      "put": {
        "summary": "Atualizar tutor por ID",
        "tags": ["Tutores"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do tutor"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorCreation"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Tutor atualizado"},
          "404": {"description": "Tutor não encontrado"},
          "500": {"description": "Erro ao atualizar tutor"},
          "delete": {
            "summary": "Deletar tutor por ID",
            "tags": ["Tutores"],
            "parameters": [
              {
                "in": "path",
                "name": "id",
                "required": true,
                "schema": {"type": "integer"},
                "description": "ID do tutor"
              }
            ],
            "responses": {
              "200": {"description": "Tutor deletado"},
              "404": {"description": "Tutor não encontrado"},
              "500": {"description": "Erro ao deletar tutor"}
            }
          }
        }
      },
      "patch": {
        "summary": "Atualizar tutor por ID (parcial)",
        "tags": ["Tutores"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do tutor"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorCreationPartial"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Tutor atualizado parcialmente com sucesso"},
          "404": {"description": "Tutor não encontrado"},
          "500": {"description": "Erro ao atualizar o tutor parcialmente"}
        }
      }
    },
    "/tutors/{id}": {
      "get": {
        "summary": "Buscar tutor por ID",
        "tags": ["Tutores"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do tutor"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do tutor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tutor"
                }
              }
            }
          },
          "404": {"description": "Tutor não encontrado"},
          "500": {"description": "Erro ao buscar tutor"}
        }
      },
      "put": {
        "summary": "Atualizar tutor por ID",
        "tags": ["Tutores"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do tutor"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorCreation"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Tutor atualizado"},
          "404": {"description": "Tutor não encontrado"},
          "500": {"description": "Erro ao atualizar tutor"}
        }
      },
      "delete": {
        "summary": "Deletar tutor por ID",
        "tags": ["Tutores"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {"type": "integer"},
            "description": "ID do tutor"
          }
        ],
        "responses": {
          "200": {"description": "Tutor deletado"},
          "404": {"description": "Tutor não encontrado"},
          "500": {"description": "Erro ao deletar tutor"}
        }
      }
    }
  }
}